requirements.txt
# Web Framework
flask

# Web Scraping & Automation
playwright

# AI / Machine Learning / Image Processing
torch
torchvision
transformers
Pillow

# Data Handling (Optional but recommended)
requests